full map data structure
